cleanup: pull stats into separate component, style cleanup
